feat(about): move selected-skill detail below hobbies
motivation:
---
- The about page previously rendered the selected-skill detail panel
  (name, subheading, description, experience tags) directly above the
  skill bars grid, both inside a single SkillSelector island under
  `cat skills.json`.
- The redesign moves the reactive detail view to the very bottom of the
  page, below `cat hobbies.json`, while keeping the bars grid in place
  near the top.

approach:
---
- Split the old `SkillSelector.tsx` into two components: `SkillBars.tsx`
  (renders only the clickable skill bars grid) and `SkillDetail.tsx`
  (renders only the name/subheading/description/experience-tags detail).
  Both are mounted as separate `client:visible` islands in about.astro.
- Since Solid context does not cross Astro island boundaries, state is
  shared via a new `src/components/about/skill-store.ts` module exporting
  a module-level `createSignal<Skill>` pair (`selected_skill` /
  `set_selected_skill`), defaulting to the first skill in
  `src/assets/skills.ts`. Astro/Vite bundles same-page islands importing
  the same source module into one shared chunk, so both islands observe
  the same signal instance — confirmed in the build output, where
  `skill-store.*.js` is emitted once and imported by both
  `SkillBars.*.js` and `SkillDetail.*.js`.
- `SkillCard.tsx` (bars + subheading) is reused as-is by both islands —
  no rewrite.
- about.astro: added a new trailing section (`cat skills/selected.md`
  PromptLine + `<SkillDetail>`) after the hobbies section; the bars
  section under `cat skills.json` is otherwise unchanged.
- Updated AGENTS.md with the shared-signal-module pattern (for reuse on
  any future cross-island state need) and fixed stale `SkillSelector`
  references left over from earlier phases.

testing:
---
- bun run typecheck / lint / fmt:check / test all green after the split
  (fixed two new lint errors: an unnecessary `as Skill[]` type assertion
  in skill-store.ts, and an always-truthy conditional in SkillDetail.tsx
  now that indexed lookup is statically known non-optional).
- bun run build confirms the shared `skill-store` chunk is emitted once
  and imported by both islands.
- Interactive Playwright check on the built site: default load shows the
  first skill in the bottom detail panel; clicking a different bar in
  the top section updates the bottom panel's text live, confirmed by
  comparing before/after text content (not just a visual diff).
